Agregando page register
