+++ client/src/components/register/Register.css
@@ -1,0 +1,36 @@
+/* Estilos para la página de registro */
+.register-container {
+    max-width: 400px;
+    margin: 0 auto;
+    padding: 20px;
+    border: 1px solid #ddd;
+    border-radius: 5px;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
+    background-color: #f5efe7;
+  }
+  
+  .form-group {
+    margin-bottom: 15px;
+  }
+  
+  label {
+    font-weight: bold;
+    color: #4f709c;
+  }
+  
+  .form-control {
+    border: 1px solid #d8c4b6;
+  }
+  
+  .btn-primary {
+    background-color: #4f709c;
+    border-color: #4f709c;
+  }
+  
+  .btn-primary:hover {
+    background-color: #213555;
+    border-color: #213555;
+  }
+  
+  /* Otros estilos si es necesario */
+
